Use custom assert statement in some places
Anki removes all asserts even in plugins, so if we call some function
that will change the behavior of the addon, it won't just be called and
everything will blow up. I also do not use custom assert everywhere,
because `mypy` do not respect it. It just ignores isinstance checks.